tests(test_zoneinfo_as_timezone): skip if tzinfo is not available on windows
On Windows, zoneinfo relies on tzinfo and without it the tests fail with
import error.

Also, rewrite pickle file loading.
Previously `open` was using a relative path, which means the path
used to depend on the current working directory and could fail if
tests were triggered from a wrong directory.
The new load_pickle function will handle that.